tests/wpt/web-platform-tests/css/css-syntax/cdc-vs-ident-tokens.html

<!doctype html>
<title>CDC versus Ident Token</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

-->

--foo { color: blue; }

</style>

<meta name=author content="Tab Atkins-Bittner">
<link rel=help href="https://drafts.csswg.org/css-syntax/#consume-token">

<!--
The ordering of the checks in the HYPHEN-MINUS step is important;
if you get it wrong, ident-token can swallow cdc-token.
-->

<script>

test(()=>{
    const rule = document.styleSheets[0].cssRules[0];
    assert_equals(rule.selectorText, "--foo");
}, "CDC-token is properly emitted, and not parsed as an ident.");

</script>
