LayoutTests/imported/w3c/web-platform-tests/css/css-syntax/cdc-vs-ident-tokens-expected.txt


PASS CDC-token is properly emitted, and not parsed as an ident.
